backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
GeoAlchemy2>=0.14.3
alembic>=1.13.1
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
reportlab>=4.0.9
numpy>=1.26.0
aiofiles>=23.2.1

# Optional — install separately if needed:
# celery[redis]>=5.3.6
# redis>=5.0.1
# scipy>=1.12.0
# matplotlib>=3.8.0
# SimPEG>=0.21.0
# pygimli>=1.5.0
# meshio>=5.3.0
# discretize>=0.10.0
